railo-cloud/requirements.cloud.txt
-r requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
redis==5.2.1
rq==1.16.1
SQLAlchemy==2.0.35
psycopg[binary]==3.1.18
pydantic-settings==2.7.0
alembic==1.14.0
httpx==0.27.2
slowapi==0.1.9
PyJWT==2.10.1
